fix(system): the ripple ignored touches on a component's own children
Pressing a button's padding rippled; pressing its label did nothing. The
overlay carried the touch handlers, and it is an absolute-fill *sibling* of the
component's children rather than their parent — touches on the label bubble up
to the Pressable and never reach it.

So the root drives the ripple and the overlay only draws it. The root is the
touch surface: everything inside it bubbles there, padding and label alike. The
overlay is pointerEvents=none and reads the two waves from context, which also
removes any chance of it interfering with the press underneath.

This is the second half of the same lesson as the first fix, and the two look
contradictory unless both are stated: the handlers cannot sit on the overlay
because it does not hear the children, and the feedback context cannot sit
inside the root because a Slot would swallow it. pressable-feedback.md now says
which and why.
