+++ packages/native/src/system/pressable-feedback/index.ts
@@ -35,5 +35,6 @@ export type {
   FeedbackVariant,
   PressableFeedbackProps,
   ResolvedAnimation,
+  RippleWave,
   SlotAnimation,
 } from './pressable-feedback.type'
