+++ packages/native/src/system/pressable-feedback/pressable-feedback-ripple.tsx
@@ -1,33 +1,16 @@
-import { useRef, useState } from 'react'
+import { useState } from 'react'
 import { StyleSheet, View } from 'react-native'
-import type {
-  GestureResponderEvent,
-  LayoutChangeEvent,
-  StyleProp,
-  ViewStyle,
-} from 'react-native'
-import Animated, {
-  Easing,
-  useAnimatedStyle,
-  useSharedValue,
-  withDelay,
-  withTiming,
-} from 'react-native-reanimated'
-import type { SharedValue } from 'react-native-reanimated'
+import type { LayoutChangeEvent, StyleProp, ViewStyle } from 'react-native'
+import Animated, { useAnimatedStyle } from 'react-native-reanimated'
 import { useXAUITheme } from '../../theme/theme-hooks'
 import { useFeedback } from './pressable-feedback-context'
 import {
-  RIPPLE_CONFIRM_DURATION,
-  RIPPLE_EXPAND_DURATION,
-  RIPPLE_FADE_IN,
-  RIPPLE_FADE_OUT,
-  RIPPLE_FADE_OUT_DELAY,
   RIPPLE_OPACITY,
   RIPPLE_START_SCALE,
   resolveSlotAnimation,
   rippleRadiusFor,
 } from './pressable-feedback.animation'
-import type { SlotAnimation } from './pressable-feedback.type'
+import type { RippleWave as Wave, SlotAnimation } from './pressable-feedback.type'
 
 export type PressableFeedbackRippleProps = {
   /**
@@ -44,33 +27,26 @@ export type PressableFeedbackRippleProps = {
 /**
  * A circle washing outwards from where the finger landed.
  *
- * **It carries its own touch handlers, and that is why it works at all.** `Pressable` owns
- * the responder system — it decides whether a touch becomes a press — and swallows the raw
- * touch props handed to it, so a ripple driven from the root draws nothing. The handlers
- * belong on this overlay's own `View`, which does not claim the responder, so the press
- * underneath is untouched.
+ * **The root drives it, this only draws it.** The root is the touch surface: touches on a
+ * component's own children — a button's label — bubble up to the `Pressable`, and never to
+ * this overlay, which is their sibling rather than their parent. An overlay owning the
+ * handlers works on the padding and does nothing on the text, which is the kind of bug that
+ * looks like a rendering problem.
  *
  * The motion is Material's `InkRipple`, and the parts that matter are the ones that are not
  * obvious: the ink is **independent of the expansion**, the circle **starts at 30%** of its
  * target rather than at a point, the target is **half the diagonal**, and the centre
- * **travels** from the finger to the middle of the control. See the constants for why each
- * one is load-bearing.
- *
- * Two waves, used in turn, so a rapid double tap opens a fresh one under the one still
- * finishing instead of cutting it.
+ * **travels** from the finger to the middle of the control. The constants carry the why.
  */
 export function PressableFeedbackRipple({
   style,
   animation: override,
 }: PressableFeedbackRippleProps) {
-  const { animation } = useFeedback()
+  const { animation, waves } = useFeedback()
   const theme = useXAUITheme()
 
   const settings = resolveSlotAnimation(override, animation.ripple, RIPPLE_OPACITY)
-
   const [size, setSize] = useState({ width: 0, height: 0 })
-  const waves = [useWave(), useWave()] as const
-  const onFirst = useRef(true)
 
   const handleLayout = (event: LayoutChangeEvent) => {
     const { width, height } = event.nativeEvent.layout
@@ -81,45 +57,15 @@ export function PressableFeedbackRipple({
     )
   }
 
-  const handleTouchStart = (event: GestureResponderEvent) => {
-    const wave = onFirst.current ? waves[0] : waves[1]
-    onFirst.current = !onFirst.current
-
-    const { locationX, locationY } = event.nativeEvent
-    wave.origin.value = { x: locationX, y: locationY }
-    wave.expand.value = 0
-    wave.expand.value = withTiming(1, {
-      duration: RIPPLE_EXPAND_DURATION,
-      easing: Easing.ease,
-    })
-    wave.alpha.value = withTiming(1, { duration: RIPPLE_FADE_IN })
-  }
-
-  const handleTouchEnd = () => {
-    // The wave catches up rather than being cut: the expansion finishes fast, and the ink
-    // only starts leaving once it has arrived.
-    const wave = onFirst.current ? waves[1] : waves[0]
-    wave.expand.value = withTiming(1, {
-      duration: RIPPLE_CONFIRM_DURATION,
-      easing: Easing.ease,
-    })
-    wave.alpha.value = withDelay(
-      RIPPLE_FADE_OUT_DELAY,
-      withTiming(0, { duration: RIPPLE_FADE_OUT })
-    )
-  }
-
-  if (!settings.enabled) return null
+  if (!waves || !settings.enabled) return null
 
   const radius = rippleRadiusFor(size.width, size.height)
 
   return (
     <View
+      pointerEvents="none"
       style={[StyleSheet.absoluteFill, styles.clip]}
       onLayout={handleLayout}
-      onTouchStart={handleTouchStart}
-      onTouchEnd={handleTouchEnd}
-      onTouchCancel={handleTouchEnd}
     >
       {waves.map((wave, index) => (
         <RippleWave
@@ -137,20 +83,6 @@ export function PressableFeedbackRipple({
 }
 
 PressableFeedbackRipple.displayName = 'XAUI.PressableFeedback.Ripple'
-
-type Wave = {
-  expand: SharedValue<number>
-  alpha: SharedValue<number>
-  origin: SharedValue<{ x: number; y: number }>
-}
-
-function useWave(): Wave {
-  return {
-    expand: useSharedValue(0),
-    alpha: useSharedValue(0),
-    origin: useSharedValue({ x: 0, y: 0 }),
-  }
-}
 
 function RippleWave({
   wave,
@@ -172,9 +104,9 @@ function RippleWave({
     const t = wave.expand.value
     const from = wave.origin.value
 
-    // The circle is laid out at its target size and scaled down, so nothing re-lays out
-    // mid-wave. It never starts smaller than 30% — a wave from a dot is invisible for the
-    // part of its life where it would read as a wave.
+    // Laid out at its target size and scaled down, so nothing re-lays out mid-wave. It
+    // never starts smaller than 30% — a wave from a dot is invisible for the part of its
+    // life where it would read as a wave.
     const scale = RIPPLE_START_SCALE + (1 - RIPPLE_START_SCALE) * t
 
     // Travels from the finger to the middle of the control as it opens, which is what
@@ -191,7 +123,6 @@ function RippleWave({
 
   return (
     <Animated.View
-      pointerEvents="none"
       style={[
         styles.wave,
         {
